Code refactoring to move into a component the city zone message
